+++ common/model/instance.go
@@ -318,78 +318,6 @@ func ExpandStore2Instance(es *ExpandInstanceStore) *Instance {
 	return out
 }
 
-type NamespaceSet struct {
-	container map[string]*Namespace
-}
-
-func NewNamespaceSet() *NamespaceSet {
-	return &NamespaceSet{
-		container: make(map[string]*Namespace),
-	}
-}
-
-func (set *NamespaceSet) Add(val *Namespace) {
-	set.container[val.Name] = val
-}
-
-func (set *NamespaceSet) Remove(val *Namespace) {
-	delete(set.container, val.Name)
-}
-
-func (set *NamespaceSet) ToSlice() []*Namespace {
-	ret := make([]*Namespace, 0, len(set.container))
-
-	for _, v := range set.container {
-		ret = append(ret, v)
-	}
-
-	return ret
-}
-
-func (set *NamespaceSet) Range(fn func(val *Namespace) bool) {
-	for _, v := range set.container {
-		if !fn(v) {
-			break
-		}
-	}
-}
-
-type ServiceSet struct {
-	container map[string]*Service
-}
-
-func NewServiceSet() *ServiceSet {
-	return &ServiceSet{
-		container: make(map[string]*Service),
-	}
-}
-
-func (set *ServiceSet) Add(val *Service) {
-	set.container[val.ID] = val
-}
-
-func (set *ServiceSet) Remove(val *Service) {
-	delete(set.container, val.ID)
-}
-
-func (set *ServiceSet) ToSlice() []*Service {
-	ret := make([]*Service, 0, len(set.container))
-
-	for _, v := range set.container {
-		ret = append(ret, v)
-	}
-
-	return ret
-}
-
-func (set *ServiceSet) Range(fn func(val *Service) bool) {
-	for _, v := range set.container {
-		if !fn(v) {
-			break
-		}
-	}
-}
-
 // CreateInstanceModel 创建存储层服务实例模型
 func CreateInstanceModel(serviceID string, req *apiservice.Instance) *Instance {
 	// 默认为健康的
