+++ common/utils/funcs_test.go
@@ -17,9 +17,19 @@
 
 package utils
 
-import "testing"
+import (
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+)
 
 func TestNewUUID(t *testing.T) {
 	val := NewUUID()
 	t.Log(val)
+}
+
+func TestStringSliceDeDuplication(t *testing.T) {
+	s := []string{"1", "2", "", "_invalid", "1", "23"}
+	s2 := StringSliceDeDuplication(s)
+	assert.Equal(t, s2, []string{"1", "2", "", "_invalid", "23"})
 }
